viafb: Add support for 2D accelerated framebuffer on VX800/VX855
viafb: Add support for 2D accelerated framebuffer on VX800/VX855

The VX800/VX820 and the VX855/VX875 chipsets have a different 2D
acceleration engine called "M1".  The M1 engine has some subtle
(and some not-so-subtle) differences to the previous engines,
so support for accelerated framebuffer on those chipsets was
disabled so far.

This patch adds full support for the M1 engine to viafb.

Signed-off-by: Harald Welte <HaraldWelte@viatech.com>